# DMS_api_gateway/docs/swagger.yaml
basePath: /
definitions:
...
  entity.Contragent:
    properties:
      address:
        type: string
      id:
        type: integer
      inn_kpp:
        maxLength: 20
        minLength: 20
        type: string
      name:
        type: string
    required:
    - address
    - inn_kpp
    - name
    type: object
  entity.Dispetcher:
    properties:
      full_name:
        type: string
      id:
        type: integer
    required:
    - full_name
    type: object
...
  entity.UpdateContragentInput:
    properties:
      address:
        type: string
      inn_kpp:
        type: string
      name:
        type: string
    type: object
  entity.UpdateDispetcherInput:
    properties:
      full_name:
        type: string
    type: object
...
  handler.errorResponse:
    properties:
      message:
        type: string
    type: object
...
  handler.getContragentsResponse:
    properties:
      contragents:
        items:
          $ref: '#/definitions/entity.Contragent'
        type: array
    type: object
  handler.getDispetchersResponse:
    properties:
      dispetchers:
        items:
          $ref: '#/definitions/entity.Dispetcher'
        type: array
    type: object
...
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for Document Management Service
  title: DMS API
  version: "1.0"
paths:
...
  /contragents:
    get:
      consumes:
      - application/json
      description: gets data about all contragents
      operationId: get-contragents
      produces:
      - application/json
      responses:
        "200":
          description: contragents info
          schema:
            $ref: '#/definitions/handler.getContragentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Contragents
      tags:
      - contragents
    post:
      consumes:
      - application/json
      description: creates data about the contragent
      operationId: create-contragent
      parameters:
      - description: contragent info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.Contragent'
      produces:
      - application/json
      responses:
        "200":
          description: contragent ID
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Contragent
      tags:
      - contragents
  /contragents/{contragent_id}:
    delete:
      consumes:
      - application/json
      description: deletes data about contragent
      operationId: delete-contragent
      parameters:
      - description: contragent id
        in: path
        name: contragent_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Contragent
      tags:
      - contragents
    put:
      consumes:
      - application/json
      description: updates data about the contragent
      operationId: update-contragent
      parameters:
      - description: contragent id
        in: body
        name: contragent_id
        required: true
        schema:
          type: integer
      - description: update contragent info
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateContragentInput'
      produces:
      - application/json
      responses:
        "200":
          description: updated contragent
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Contragent
      tags:
      - contragents
  /dispetchers:
    get:
      consumes:
      - application/json
      description: gets data about all dispetchers
      operationId: get-dispetchers
      produces:
      - application/json
      responses:
        "200":
          description: dispetchers info
          schema:
            $ref: '#/definitions/handler.getDispetchersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Dispetchers
      tags:
      - dispetchers
    post:
      consumes:
      - application/json
      description: creates and adds data about the dispetcher
      operationId: create-dispetcher
      parameters:
      - description: dispetcher info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/entity.Dispetcher'
      produces:
      - application/json
      responses:
        "200":
          description: dispetcher ID
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Dispetcher
      tags:
      - dispetchers
  /dispetchers/{dispetcher_id}:
    delete:
      consumes:
      - application/json
      description: deletes data about dispetcher
      operationId: delete-dispetcher
      parameters:
      - description: dispetcher id
        in: body
        name: dispetcher_id
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Dispetcher
      tags:
      - dispetchers
    put:
      consumes:
      - application/json
      description: updates data about the dispetcher
      operationId: update-dispetcher
      parameters:
      - description: dispetcher id
        in: body
        name: dispetcher_id
        required: true
        schema:
          type: integer
      - description: update dispetcher info
        in: body
        name: updateData
        required: true
        schema:
          $ref: '#/definitions/entity.UpdateDispetcherInput'
      produces:
      - application/json
      responses:
        "200":
          description: updated dispetcher
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update Dispetcher
      tags:
      - dispetchers
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"